wip: clean up and refactoring
