+++ cmds/modules/networkd/addr_watcher.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/containernetworking/plugins/pkg/ns"
 	"github.com/rs/zerolog/log"
 
 	"time"
@@ -16,10 +15,8 @@ import (
 	"github.com/threefoldtech/tfexplorer/schema"
 	"github.com/threefoldtech/zos/pkg"
 	"github.com/threefoldtech/zos/pkg/network/ifaceutil"
-	"github.com/threefoldtech/zos/pkg/network/namespace"
 	"github.com/threefoldtech/zos/pkg/network/ndmz"
 	"github.com/threefoldtech/zos/pkg/network/types"
-	"github.com/threefoldtech/zos/pkg/network/yggdrasil"
 	"github.com/vishvananda/netlink"
 )
 
@@ -27,9 +24,10 @@ type WatchedLinks struct {
 	linkNames map[string]struct{}
 	dir       client.Directory
 	nodeID    pkg.Identifier
+	dmz       ndmz.DMZ
 }
 
-func NewWatchedLinks(linkNames []string, nodeID pkg.Identifier, dir client.Directory) WatchedLinks {
+func NewWatchedLinks(dmz ndmz.DMZ, linkNames []string, nodeID pkg.Identifier, dir client.Directory) WatchedLinks {
 	names := make(map[string]struct{}, len(linkNames))
 
 	for _, n := range linkNames {
@@ -40,6 +38,7 @@ func NewWatchedLinks(linkNames []string, nodeID pkg.Identifier, dir client.Direc
 		linkNames: names,
 		dir:       dir,
 		nodeID:    nodeID,
+		dmz:       dmz,
 	}
 }
 
@@ -61,7 +60,7 @@ func (w WatchedLinks) callBack(update netlink.AddrUpdate) error {
 		return err
 	}
 
-	ndmzIfaces, err := getNdmzInterfaces()
+	ndmzIfaces, err := w.dmz.Interfaces()
 	if err != nil {
 		return err
 	}
@@ -106,54 +105,6 @@ func (w WatchedLinks) Forever(ctx context.Context) error {
 			}
 		}
 	}
-}
-
-func getNdmzInterfaces() ([]types.IfaceInfo, error) {
-	var output []types.IfaceInfo
-
-	f := func(_ ns.NetNS) error {
-		links, err := netlink.LinkList()
-		if err != nil {
-			log.Error().Err(err).Msgf("failed to list interfaces")
-			return err
-		}
-		for _, link := range links {
-			name := link.Attrs().Name
-			if name == ndmz.DMZPub4 || name == ndmz.DMZPub6 || name == yggdrasil.YggIface {
-				addrs, err := netlink.AddrList(link, netlink.FAMILY_ALL)
-				if err != nil {
-					return err
-				}
-
-				info := types.IfaceInfo{
-					Name:       name,
-					Addrs:      make([]types.IPNet, len(addrs)),
-					MacAddress: schema.MacAddress{link.Attrs().HardwareAddr},
-				}
-				for i, addr := range addrs {
-					info.Addrs[i] = types.NewIPNet(addr.IPNet)
-				}
-
-				output = append(output, info)
-			}
-
-		}
-		return nil
-	}
-
-	// get the ndmz network namespace
-	ndmz, err := namespace.GetByName(ndmz.NetNSNDMZ)
-	if err != nil {
-		return nil, err
-	}
-	defer ndmz.Close()
-
-	err = ndmz.Do(f)
-	if err != nil {
-		return nil, err
-	}
-
-	return output, nil
 }
 
 func getLocalInterfaces() ([]types.IfaceInfo, error) {
